계산기

+++ LikeLion/src/Test_20221005/Calculator.java
@@ -4,5 +4,5 @@ public class Calculator {
     public int add(int a, int b) { return a+b; }
     public int minus(int a, int b) { return a-b; }
     public int multi(int a, int b) { return a*b; }
-    public int divide(int a, int b) { return a/b; }
+    public double divide(int a, int b) { return (double)a/(double)b; }
 }
